lite: gitstatus plugin enhancements
first of all, it now supports showing the modified lines
via the diff parsing from gitdiff's parser. the stat is
now the 1st thing on the right part of the statusview,
and i changed some of the style variables.

the plugin now *only* changes the info when the docview
is changed, instead of all the time while in the docview