awesome(syntax): refactors/improvements
icons in the syntax theme now use svgs instead
of fonts. this makes them easier to deal with
because i couldnt even get the power menu icons
to work when i reinstalled my setup

the systray has a better looking icon (well,
it has an actual icon at all) now

added a generic button and icon function to use
the svgs. in the future itll be used in the stardew
theme

music control buttons have been improved due to the
new button widget and now change cursor on hover and
have a circular background which shows up on hover