awesome(syntax): add caps lock indicator, fix popup autohide with animations
previously, if the noAnimate setting is false, when clicking anywhere
to hide popups (like the start menu) it would hide it and remove the
ability for it to come back. that's been fixed.

added an indicator for when caps lock is toggled since my new laptop
doesnt have an led for it